+++ packages/core/src/common/theme.ts
@@ -51,6 +51,7 @@ export const shadows = [
  * Generic padding styles by size
  */
 export const padding = {
+  xs: "0.1rem 0.2rem",
   sm: "0.15rem 0.3rem",
   md: "0.35rem 0.4rem",
   lg: "0.45rem 0.6rem"